+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 
-# 主要
+# TODO
 
-翻譯
-發音
+- beautify css (style and class)
+- add and unit tests
 
-highlight
-git
+- 翻譯
+- 發音
+
+# DONE
+
+- highlight text
+- delete sentence
+  - confirm
+
+# CHANGE LOG
+
+- 20210310
+  1. improve interface
+  2. add yes-no-dialog and ask user to confirm deletion